Splits function tests for each DB into dedicated modules
